Removed the "create ticket on" options
It may cause some weird issues when creating a ticket (who created it, message, custom fields, etc), so pulling this out for now. There will be a way to create a ticket from an already created event via a query string that is passed to the admin_new_ticket page.